Changed the Blast Client : can blast on Uniprot, Swissprot and Intact using only a sequence.
Added a new Class BlastProtein which contains information such as AC, score, identity, description for each hit of a blast.

Modified BlastInput and Job to support blasts with a sequence and without uniprot AC.

Added a class BlastResultFilter to be able to parse the results and keep only the hits with a certain identity and/or organism

Added  new RunTimeException BlastResultFilterException.

Added a new class ProteinWsWuBlastService to only use the new methods blastClient

Changed the Blast Client Test

Added BlastResultFilterTest and ProteinWsWuBlastServiceTest.

Changed the pom file to add a dependency on intact-uniprot