Added a new event : RangeOutOfBoundEvent to manage range which are not in the sequence
Added a new Method in ProteinUpdateProcessor, AbstractProteinUpdateProcessorListener and ProteinUpdateProcessorListener to manage out of bound ranges.

Fixed a new bug in DuplicateFixer and XRefUpdaterUtils : the intact-secondary was deleted because not in uniprot. Changed one test in consequence : FileReportHandlerTest

Now we are writing a full report on out of bound ranges : changed LoggingProcessorListener, ReportWriterListener, FileReportHandler and UpdateReportHandler

Added a new class OutOfBoundRange to store the information about the range.

Changed RangeChecker and RangeFixer : now we are shifting the ranges only if it is not out of bound.